moved download zip to utils
